LayoutTests/platform/glib/imported/w3c/web-platform-tests/websockets/remove-own-iframe-during-onerror.window-expected.txt

CONSOLE MESSAGE: WebSocket connection to 'ws://web-platform.test:49001/does-not-exist' failed: The server did not accept the WebSocket handshake.

PASS removing an iframe from within an onerror handler should work
